Data flow from App - Props
